+++ CHANGELOG.md
@@ -2,6 +2,8 @@
   - Depend on logstash-mixin-zeromq, remove lib/logstash/util/zeromq
   - Adapt test usage to be friendly with the way we run LS core default plugins test
   - Fix plugin consumes 100% cpu
+  - Properly set topic field in pubsub topology
+  - Support multiple message frames
 # 2.0.4
   - Depend on logstash-core-plugin-api instead of logstash-core, removing the need to mass update plugins on major releases of logstash
 # 2.0.3